+++ src/pageEditor/tabs/effect/BlockConfiguration.tsx
@@ -19,7 +19,7 @@ import React, { useMemo, useRef } from "react";
 import { RegistryId } from "@/core";
 import { getIn, useFormikContext } from "formik";
 import useBlockOptions from "@/hooks/useBlockOptions";
-import { Button, Card } from "react-bootstrap";
+import { Card } from "react-bootstrap";
 import SchemaFieldContext from "@/components/fields/schemaFields/SchemaFieldContext";
 import devtoolFieldOverrides from "@/pageEditor/fields/devtoolFieldOverrides";
 import Loader from "@/components/Loader";
@@ -38,8 +38,7 @@ import FieldSection from "@/pageEditor/fields/FieldSection";
 import getType from "@/runtime/getType";
 import { FormState } from "@/pageEditor/pageEditorTypes";
 import { useGetMarketplaceListingsQuery } from "@/services/api";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons";
+import ConfigurationTitle from "./ConfigurationTitle";
 
 const rootModeOptions = [
   { label: "Inherit", value: "inherit" },
@@ -68,33 +67,6 @@ const BlockConfiguration: React.FunctionComponent<{
   const [{ block, error }, BlockOptions] = useBlockOptions(blockId);
   const { data: listings = {} } = useGetMarketplaceListingsQuery();
   const listing = listings[blockId];
-
-  console.log("configuration", {
-    blockId,
-    block,
-    listing,
-  });
-  const configurationTitle = (
-    <span>
-      Input: <span className="text-muted">{block?.name}</span>
-    </span>
-  );
-  const configurationHeader =
-    (listing?.instructions != null && listing?.instructions != "") ||
-    listing?.assets?.length > 0 ? (
-      <div className="d-flex justify-content-between">
-        {configurationTitle}
-        <a
-          href={`https://www.pixiebrix.com/marketplace/${listing.id}/`}
-          target="_blank"
-          rel="noreferrer"
-        >
-          <FontAwesomeIcon icon={faExternalLinkAlt} /> View Documentation
-        </a>
-      </div>
-    ) : (
-      configurationTitle
-    );
 
   // Conditionally show Advanced options "Condition" and "Target" depending on the value of blockType.
   // If blockType is undefined, don't show the options.
@@ -136,7 +108,9 @@ const BlockConfiguration: React.FunctionComponent<{
       <AdvancedLinks name={name} scrollToRef={advancedOptionsRef} />
 
       <Card>
-        <FieldSection title={configurationHeader}>
+        <FieldSection
+          title={<ConfigurationTitle block={block} listing={listing} />}
+        >
           <SchemaFieldContext.Provider value={devtoolFieldOverrides}>
             {blockErrors?.id && (
               <div className="invalid-feedback d-block mb-4">
